module_decls
